refactor/#125: UserConnectionDisconnectedDomainEvent를 user의 도메인 이벤트 폴더로 이동
