+++ src/features/user/domain/events/user-connection-disconnected.domain-event.ts
