+++ src/features/user/domain/user.entity.ts
@@ -29,7 +29,7 @@ import { HttpUnprocessableEntityException } from '@libs/exceptions/client-errors
 import { AggregateID } from '@libs/ddd/entity.base';
 import { HttpForbiddenException } from '@libs/exceptions/client-errors/exceptions/http-forbidden.exception';
 import { UserMbtiUnion } from '@features/user/types/user.type';
-import { UserConnectionDisconnectedDomainEvent } from '@features/user/user-connection/domain/events/user-connection-disconnected.domain-event';
+import { UserConnectionDisconnectedDomainEvent } from '@features/user/domain/events/user-connection-disconnected.domain-event';
 
 export class UserEntity extends AggregateRoot<UserProps> {
   static readonly USER_ATTACHMENT_URL = process.env.USER_ATTACHMENT_URL;
